refactor(block): renamed utils file for drives
Renamed file with utils for vhost user block
so it would be able to contain utils for both
virtio and vhost-user drives.
